Reformats button rows and created components for reusability/cleanup
